[ENGOPS-411] Fine grain control of shims dependencies Added a version property for each of the shims. If necessary, these will be updated at build-time for release builds of the plugin.
**** NOTICE *******************************
New properties should be added for new shims as needed and any ticket for this effort should have a sub-task for ENGOPS to do the matching work in the globlal version properties file(s).
Old properties should be removed as needed and any ticket for this effort should have a sub-task for ENGOPS to do the matching work in the globlal version properties file(s).
**********************************************