[PDI-13675]: Saving HBase mapping in HBaseRowDecoder step causes error
There is a situation when  the step contains the mapping without any name.
The cause of this issue is a such case.
In this case we get no one output fields in RowMeta for this step (method HBaseRowDecoderMeta.getFields())
And then during verification TransMeta (method JobEntryHadoopTransJobExecutor.verifyTransMeta()) we throw KettleException on checking meta of Hadoop Output step.
So the fix is to allow getting fields from mapping without name (so-called "local" mapping).
Also added a couple of unit tests.