[BACKLOG-5230] - REGRESSION: HBase Input fails with "null key value" error message
the issue was caused by changes for new HBase filters.
in order to implement key filtering the corresponding column was added to the alias-map
however, as soon as the "key" is a synthetic column it was not available by the alias.

Solution is to skip the "key" alias resolution when looping through columns' map.