[BAD-341]: Hadoop File Input: NPE appears after click on "Browse" button
The problem was caused with the case when the user types any nonsensical sequence of symbols in Path filed.
In such case defaultInitialFile=null an we have NPE.
Just Fixed NPEs without any changes in the logic of working on widgetSelected.
No unit tests has been added because these changes  are in UI components.