[BACKLOG-5908]: REG: Hadoop Job Executer doesn't use correct information from named cluster.
This issue was introduced with changes in the commit: https://github.com/pentaho/big-data-plugin/commit/ceb855aa6b215b43a82f9b5c353527a29b83d3d7
(exactly starting from the line:
https://github.com/pentaho/big-data-plugin/blob/ceb855aa6b215b43a82f9b5c353527a29b83d3d7/kettle-plugins/mapreduce/src/main/java/org/pentaho/big/data/kettle/plugins/mapreduce/JobEntryHadoopJobExecutorDialog.java#L228)

So the fix is just to return the behavior that was before.
No unit test because the change is on UI (in Dialog).