[BACKLOG-9186]: Exception `invalid constant type` appears in PMR mapper logs on cluster
The problem was:
During configuration the map runner is creating a new tranformation.
The process creation of the transformation involves calling at org.pentaho.di.core.KettleEnvironment.init.
And then we are trying to register plugins and types: PluginRegistry.init(), BasePluginType.registerPluginJars();
We use scannotation to find all annotated classes in jars but scannotation uses javassist for this.

So if any of classes in jar contains Java 8 lambdas, we will have this issue.
This occurs because Java 8 lambdas uses invokedynamic to generate code in runtime but javassist 3.12.1.GA doesn't support it.
With the fix https://issues.jboss.org/browse/JASSIST-174 javassist has got support for the CONSTANT_InvokeDynamic=18
So the fix is just to upgrade javassist to the latest release 3.20.0-GA.