[BACKLOG-10984] ClusterInitiazation during driver load causes sys init to hang
Removing some logic which determined the named cluster to
associated with a jdbc connection by looking at the Url.
That association is currently unnecessary since we are not yet
supporting jdbc connections within named clusters.  The logic
was creating circularity during init:  the ClusterInitializingDriver
could be called when JackRabbit makes a repository connection.
But determining named cluster in turn requires a Repository.

We'll need to rethink this design when we have functional requirements
to associate named clusters and JDBC connections.