fix[SP-6856]: remove test references to deprecated use of userinfo
VFS2 2.10.0 is not producing URIs as 2.8.0 did in respect of the userinfo component.

From the testing done so far, this does not seem to have impact on the solution.
However, for now, I'm @Ignoring the tests referring this functionality, in case we need to reinstate it later.

(cherry picked from commit 0fa81d67c167396f582323124187459a83e97d0e)